Fix unwanted refreshes when setting size of wxChoice in wxMSW
The window was still resized even if its size didn't really change,
which, in particular, unexpectedly resulted in closing its drop down if
it was shown.

Fix this by moving the code dealing with the different interpretation of
the control height between wxMSW and MSW (for the former, it's just the
height of the visible part, while for the latter it includes drop down
as well), from wxChoice::DoSetSize() to DoMoveWindow(). This allows the
base class DoSetSize() to correctly filter out calls not really changing
the size, while still setting the height correctly when it does change.

See #17075.

See https://github.com/wxWidgets/wxWidgets/pull/389

Conflicts:
	src/msw/choice.cpp